rename group as controller
