+++ src/spu/src/replication/follower/mod.rs
@@ -1,7 +1,7 @@
 mod state;
 mod api_key;
 mod peer_api;
-mod group;
+mod controller;
 pub mod sync;
 
 pub use self::state::{FollowersState, SharedFollowersState, FollowerReplicaState};
